+++ src/parser/block/extension/placeholder.test.ts
@@ -8,12 +8,13 @@ describe('Unit: parser/block/extension/placeholder', () => {
 
     it('warning', () => {
       assert.deepStrictEqual(inspect(parser('~~~\n~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
+      assert.deepStrictEqual(inspect(parser('~~~\n~~~\n')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
       assert.deepStrictEqual(inspect(parser('~~~\n<"\n~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], ""]);
       assert.deepStrictEqual(inspect(parser('~~~a\n~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
       assert.deepStrictEqual(inspect(parser('~~~a \n~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
       assert.deepStrictEqual(inspect(parser('~~~a b \n~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
       assert.deepStrictEqual(inspect(parser('~~~ a\n~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~\n~~~\n')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
+      assert.deepStrictEqual(inspect(parser('~~~~\n~~~~')), [["<p><strong>WARNING: DON'T USE <code data-src=\"`~~~`\">~~~</code> SYNTAX!!</strong><br>This <em>extension syntax</em> is reserved for extensibility.</p>"], '']);
     });
 
   });