+++ src/parser/block/math.test.ts
@@ -23,6 +23,7 @@ describe('Unit: parser/block/math', () => {
 
     it('ab', () => {
       assert.deepStrictEqual(inspect(parser('$$\na\n$$')), [['<div class="math">$$\na\n$$</div>'], '']);
+      assert.deepStrictEqual(inspect(parser('$$\na\n$$\n')), [['<div class="math">$$\na\n$$</div>'], '']);
       assert.deepStrictEqual(inspect(parser('$$\na\nb\n$$')), [['<div class="math">$$\na\nb\n$$</div>'], '']);
       assert.deepStrictEqual(inspect(parser('$$\n\\\n$$')), [['<div class="math">$$\n\\\n$$</div>'], '']);
       assert.deepStrictEqual(inspect(parser('$$\n$\n$$')), [['<div class="math">$$\n$\n$$</div>'], '']);