+++ src/parser/inline.ts
@@ -54,3 +54,6 @@ export const inline: InlineParser = combine<InlineParser>([
 ]);
 
 export { index } from './inline/extension/index';
+export { label } from './inline/extension/label';
+export { media } from './inline/media';
+export { url } from './inline/autolink/url';