+++ src/parser/string/label.ts
@@ -1,0 +1,4 @@
+﻿export function makeLabel(text: string): string {
+  assert(!text.includes('\n'));
+  return `label:${text}`;
+}